Rename "manual" to api
