+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/deserializers/ObservedSpanDeserializer.java
@@ -25,7 +25,7 @@ import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
-import com.uber.jaeger.crossdock.tracetest_manual.ObservedSpan;
+import com.uber.jaeger.crossdock.api.ObservedSpan;
 
 import java.io.IOException;
