+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/deserializers/StartTraceRequestDeserializer.java
@@ -26,8 +26,8 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.uber.jaeger.crossdock.tracetest_manual.Downstream;
-import com.uber.jaeger.crossdock.tracetest_manual.StartTraceRequest;
+import com.uber.jaeger.crossdock.api.Downstream;
+import com.uber.jaeger.crossdock.api.StartTraceRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
