+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/resources/behavior/TraceBehavior.java
@@ -29,10 +29,10 @@ import com.uber.jaeger.crossdock.Constants;
 import com.uber.jaeger.crossdock.JerseyServer;
 import com.uber.jaeger.crossdock.resources.behavior.tchannel.TChannelServer;
 import com.uber.jaeger.crossdock.thrift.TracedService;
-import com.uber.jaeger.crossdock.tracetest_manual.Downstream;
-import com.uber.jaeger.crossdock.tracetest_manual.JoinTraceRequest;
-import com.uber.jaeger.crossdock.tracetest_manual.ObservedSpan;
-import com.uber.jaeger.crossdock.tracetest_manual.TraceResponse;
+import com.uber.jaeger.crossdock.api.Downstream;
+import com.uber.jaeger.crossdock.api.JoinTraceRequest;
+import com.uber.jaeger.crossdock.api.ObservedSpan;
+import com.uber.jaeger.crossdock.api.TraceResponse;
 import com.uber.tchannel.api.SubChannel;
 import com.uber.tchannel.api.TFuture;
 import com.uber.tchannel.messages.ThriftRequest;
