+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/resources/behavior/http/TraceBehaviorResource.java
@@ -26,9 +26,9 @@ import com.uber.jaeger.Span;
 import com.uber.jaeger.context.TracingUtils;
 import com.uber.jaeger.crossdock.Constants;
 import com.uber.jaeger.crossdock.resources.behavior.TraceBehavior;
-import com.uber.jaeger.crossdock.tracetest_manual.JoinTraceRequest;
-import com.uber.jaeger.crossdock.tracetest_manual.StartTraceRequest;
-import com.uber.jaeger.crossdock.tracetest_manual.TraceResponse;
+import com.uber.jaeger.crossdock.api.JoinTraceRequest;
+import com.uber.jaeger.crossdock.api.StartTraceRequest;
+import com.uber.jaeger.crossdock.api.TraceResponse;
 import io.opentracing.tag.Tags;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
