+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/resources/behavior/tchannel/JoinTraceThriftHandler.java
@@ -24,8 +24,8 @@ package com.uber.jaeger.crossdock.resources.behavior.tchannel;
 import com.uber.jaeger.crossdock.resources.behavior.TraceBehavior;
 import com.uber.jaeger.crossdock.thrift.JoinTraceRequest;
 import com.uber.jaeger.crossdock.thrift.TracedService;
-import com.uber.jaeger.crossdock.tracetest_manual.Downstream;
-import com.uber.jaeger.crossdock.tracetest_manual.TraceResponse;
+import com.uber.jaeger.crossdock.api.Downstream;
+import com.uber.jaeger.crossdock.api.TraceResponse;
 import com.uber.tchannel.api.handlers.ThriftRequestHandler;
 import com.uber.tchannel.messages.ThriftRequest;
 import com.uber.tchannel.messages.ThriftResponse;
