+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/serializers/DownstreamSerializer.java
@@ -26,7 +26,7 @@ import java.io.IOException;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
-import com.uber.jaeger.crossdock.tracetest_manual.Downstream;
+import com.uber.jaeger.crossdock.api.Downstream;
 
 public class DownstreamSerializer extends JsonSerializer<Downstream> {
