+++ jaeger-crossdock/src/main/java/com/uber/jaeger/crossdock/serializers/TraceResponseSerializer.java
@@ -26,8 +26,8 @@ import java.io.IOException;
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
-import com.uber.jaeger.crossdock.tracetest_manual.ObservedSpan;
-import com.uber.jaeger.crossdock.tracetest_manual.TraceResponse;
+import com.uber.jaeger.crossdock.api.ObservedSpan;
+import com.uber.jaeger.crossdock.api.TraceResponse;
 
 public class TraceResponseSerializer extends JsonSerializer<TraceResponse> {
