+++ jaeger-crossdock/src/test/java/com/uber/jaeger/crossdock/resources/behavior/http/TraceBehaviorResourceTest.java
@@ -27,11 +27,11 @@ import com.uber.jaeger.crossdock.Constants;
 import com.uber.jaeger.crossdock.JerseyServer;
 import com.uber.jaeger.crossdock.resources.behavior.TraceBehavior;
 import com.uber.jaeger.crossdock.resources.behavior.tchannel.TChannelServer;
-import com.uber.jaeger.crossdock.tracetest_manual.Downstream;
-import com.uber.jaeger.crossdock.tracetest_manual.JoinTraceRequest;
-import com.uber.jaeger.crossdock.tracetest_manual.ObservedSpan;
-import com.uber.jaeger.crossdock.tracetest_manual.StartTraceRequest;
-import com.uber.jaeger.crossdock.tracetest_manual.TraceResponse;
+import com.uber.jaeger.crossdock.api.Downstream;
+import com.uber.jaeger.crossdock.api.JoinTraceRequest;
+import com.uber.jaeger.crossdock.api.ObservedSpan;
+import com.uber.jaeger.crossdock.api.StartTraceRequest;
+import com.uber.jaeger.crossdock.api.TraceResponse;
 import io.opentracing.tag.Tags;
 import org.apache.log4j.BasicConfigurator;
 import org.junit.After;
